arreglo respawn time

+++ server/items/weapons/ak47.cpp
@@ -9,7 +9,7 @@ constexpr int MAX_SPREAD_COUNTER = -1;
 constexpr int BURST_INTERVAL = 5;
 
 Ak47::Ak47(uint8_t type, uint8_t id, float x_pos, float y_pos, uint8_t damage,
-           uint8_t range, uint8_t ammo_quantity, float recoil)
+           uint8_t range, int ammo_quantity, float recoil)
     : Weapon(type, id, x_pos, y_pos, damage, range, ammo_quantity, recoil),
       spread_counter(0), bullets_vector() {
   bullets_vector.emplace_back(AK47_BULLET, 1, 0, 0, damage, range, 0);
