+++ server/items/weapons/granada.cpp
@@ -5,7 +5,7 @@
 #include "granada.h"
 
 Granada::Granada(uint8_t type, uint8_t id, float x_pos, float y_pos,
-                 uint8_t damage, uint8_t range, uint8_t ammo_quantity,
+                 uint8_t damage, uint8_t range, int ammo_quantity,
                  float recoil)
     : Weapon(type, id, x_pos, y_pos, damage, range, ammo_quantity, recoil),
       bala(GRANADA_BULLET, 1, 0, 0, 10, range, 0.3, GRANADA_TIME_TO_EXPLODE),
